renamed upgrade from v4.4.2 to v4.5.0
